+++ src/pages/index.jsx
@@ -1,0 +1,40 @@
+// import React from 'react';
+import Head from 'next/head';
+import Image from 'next/image';
+import styles from '../../styles/Home.module.css';
+
+export default function Home() {
+  return (
+    <div className={styles.container}>
+      <Head>
+        <title>Learn</title>
+        <meta name="description" content="Learn with Breatheco.de" />
+        <link rel="icon" href="/favicon.ico" />
+      </Head>
+
+      <main className={styles.main}>
+        <h1 className={styles.title}>
+          Welcome to
+          {' '}
+          <a href="https://nextjs.org">Next.js!</a>
+        </h1>
+
+        <div className={styles.grid}>
+          <a href="/" className={styles.card}>
+            <h2>Hola mundo</h2>
+          </a>
+        </div>
+      </main>
+
+      <footer className={styles.footer}>
+        <a href="/">
+          Powered by
+          {' '}
+          <span className={styles.logo}>
+            <Image src="/vercel.svg" alt="Vercel Logo" width={72} height={16} />
+          </span>
+        </a>
+      </footer>
+    </div>
+  );
+}
